"changed filter"

+++ tasks/image_processor/edge_detection.h
@@ -5,7 +5,7 @@
 class EdgeDetection : public Filter {
 public:
     Image Apply(const Image& image) const override;
-    explicit EdgeDetection(int threshold);
+    explicit EdgeDetection(double threshold);
 
 private:
     double threshold_;
